requirements.txt
websocket-client==1.6.4
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
colorama==0.4.6
pymongo==4.6.1
flask==3.0.0
flask-cors==4.0.0

